pyproject.toml
[project]
name = "frontdesk-test"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "certifi>=2025.8.3",
    "flask>=3.1.2",
    "livekit-agents[openai]==1.2.11",
    "livekit-plugins-deepgram~=0.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-silero~=0.2",
    "python-dotenv>=1.1.1",
]
